requirements.txt
requests
slackclient
boto
xmltodict
urllib3==1.26.6